Moved `rest_api.py` to `ws.py`
